+++ src/main/java/com/br/apitemplate/dto/tributo/TributoResponse.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TributoResponse {
-    private int sku;
-    private int valorIcms;
-    private int valorPis;
-    private int valorDifaul;
-    private int  valorFcpIcms;
+    private Integer sku;
+    private Double valorIcms;
+    private Double valorPis;
+    private Double valorDifaul;
+    private Double  valorFcpIcms;
 }
